hmt333/requirements.txt
APScheduler==3.9.1.post1
pyserial==3.5
configparser==5.3.0
